more correct comparison is using of Naive counter with double precision vs Kahan accumulator

+++ main.cpp
@@ -64,7 +64,7 @@ public:
     std::string Name() const override;
 };
 
-using TDummyCovariationCalculator = TTypedCovariationCalculator<double>;
+using TDummyCovariationCalculator = TTypedCovariationCalculator<long double>;
 using TKahanCovariationCalculator = TTypedCovariationCalculator<TKahanAccumulator>;
 
 template <>
